Upload function - validation

+++ src/main/java/pl/pytlak/photoart/controller/PhotoController.java
@@ -1,7 +1,9 @@
 package pl.pytlak.photoart.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.tika.Tika;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -16,9 +18,12 @@ import pl.pytlak.photoart.service.photo.PhotoService;
 import javax.validation.Valid;
 import java.io.File;
 import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
+
 
 @RestController
 @RequiredArgsConstructor
@@ -29,26 +34,32 @@ public class PhotoController {
     private final AuthenticationService authenticationService;
     private final AlbumService albumService;
 
-    @PostMapping("/upload")
-    public ResponseEntity<Object> upload(@Valid @RequestBody UploadRequest uploadRequest, @RequestParam MultipartFile img) {
 
-        Path photoPath = Paths.get("C:\\Projects\\PhotoArt\\src\\main\\resources\\static\\images\\");
+    @PostMapping(value = "/upload", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
+    public ResponseEntity<Object> upload(@Valid @ModelAttribute UploadRequest uploadRequest) throws URISyntaxException, IOException {
+
+
+        if (!photoService.validationImage(uploadRequest.getImg()))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+
+        URL res = getClass().getClassLoader().getResource("static/images");
+        Path photoPath = Paths.get(res.toURI());
         User user = authenticationService.getCurrentUser();
-        String originalFilename = img.getOriginalFilename();
+        String originalFilename = uploadRequest.getImg().getOriginalFilename();
         Optional<Album> optionalAlbum = albumService.findByAlbumIdAndUserId(uploadRequest.getAlbumId(), user.getId());
 
         if (optionalAlbum.isPresent()) {
             Photo photo = Photo.builder()
-                    .name(img.getOriginalFilename())
+                    .name(uploadRequest.getImg().getOriginalFilename())
                     .title(uploadRequest.getTitle())
                     .album(optionalAlbum.get())
                     .build();
 
             Long photoId = photoService.add(photo).getId();
-            File photoFile = new File(photoPath + user.getUsername() + "-" + photoId + "-" + originalFilename);
+            File photoFile = new File(photoPath + "\\" + user.getUsername() + "-" + photoId + "-" + originalFilename);
 
             try {
-                img.transferTo(photoFile);
+                uploadRequest.getImg().transferTo(photoFile);
             } catch (IOException exception) {
                 photoService.remove(photo);
                 return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
@@ -58,5 +69,10 @@ public class PhotoController {
 
         }
         return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+
     }
+
 }
+
+
+
